# openapi.yaml
openapi: 3.0.3
info:
  title: Bambuzle API
  description: |
    REST API for the Bambuzle 3D printer monitoring dashboard.

    Bambuzle connects to BambuLab printers via MQTT, collects telemetry (temperatures, layer progress, fan speeds),
    tracks print jobs, detects anomalies, and provides real-time state via WebSocket.

    ## Authentication
    Two separate layers:
    1. **Dashboard admin token** (who may use this API). Every `/api/*` route and the WebSocket require it, except
       `/api/session`, `/api/spec` and `/api/docs` (and the `/healthz` / `/readyz` probes, which live outside `/api`). Send `Authorization: Bearer <token>`, or exchange the token for an
       HttpOnly session cookie via `POST /api/session`. `BAMBUZLE_PUBLIC_READ=true` opens GETs and the WebSocket;
       `BAMBUZLE_AUTH=off` disables the check entirely.
    2. **BambuLab Cloud login** (the server's own connection to your printers), via `/api/auth/login`.

    ## Real-time Data
    For live updates, connect to the WebSocket endpoint at `ws://<host>:<port>/ws`.
    Messages are JSON with `{ type, data }` format. Types include `state`, `event`, and `auth`.
  version: 0.7.0
  contact:
    name: Bambuzle
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/printers/{deviceId}/history:
    get:
      tags: [History]
      summary: Get temperature/progress samples
      description: Time-series telemetry samples for a printer. Defaults to most recent 5000 samples.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start of time window (ISO 8601)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End of time window (ISO 8601)
        - name: limit
          in: query
          schema:
            type: integer
            default: 5000
          description: Maximum number of samples to return
      responses:
        '200':
          description: Array of samples
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sample'
  /api/printers/{deviceId}/events:
    get:
      tags: [History]
      summary: Get events for a printer
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        '200':
          description: Array of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
  /api/events:
    get:
      tags: [History]
      summary: Get recent events across all printers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Array of events with printer names
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Event'
                    - type: object
                      properties:
                        printer_name:
                          type: string

  # ─── Jobs ───
  /api/printers/{deviceId}/jobs:
    get:
      tags: [Jobs]
      summary: Get print job history
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Array of print jobs (newest first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrintJob'
  /api/stats:
    get:
      tags: [Jobs]
      summary: Print job statistics (totals, success rates, by printer / material / day)
      description: |
        Aggregates print jobs whose `started_at` falls within [from, to]. Defaults to the last 30 days.
        Cancelled = end_state `CANCELLED` (user cancel) or legacy `IDLE`. `successRate` = finished / (finished + failed + cancelled), null when 0.
        Durations are wall-clock (pauses included); rows without `duration_sec` fall back to ended_at − started_at.
        Days in `byDay` are UTC.
      parameters:
        - name: printer
          in: query
          description: Restrict to one printer (device ID)
          schema:
            type: string
        - name: from
          in: query
          description: Window start (ISO 8601 date or date-time; offset-less values are UTC). Default now − 30 days.
          schema:
            type: string
            example: '2026-09-01'
        - name: to
          in: query
          description: Window end (ISO 8601; a bare date means the end of that UTC day). Default now.
          schema:
            type: string
            example: '2026-10-01T12:00:00Z'
      responses:
        '200':
          description: Job statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStats'
        '400':
          description: Invalid from/to (not ISO 8601, impossible date, or from after to)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/printers/{deviceId}/jobs/{jobId}/layers:
    get:
      tags: [Anomalies]
      summary: Get layer transitions for a job
      description: |
        One row per layer change detected at MQTT-message granularity.
        Includes a snapshot of temperatures, speed, and progress at each layer transition,
        plus the duration since the previous layer.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Array of layer transitions ordered by layer number
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LayerTransition'
  /api/printers/{deviceId}/jobs/{jobId}/anomalies:
    get:
      tags: [Anomalies]
      summary: Get temperature anomalies for a job
      description: |
        Only stores anomalous readings where deviation from target exceeds the configured
        threshold OR rate of change exceeds the configured threshold. Normal readings stay
        in the samples table.

        Deviation is signed: negative = undershoot, positive = overshoot.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Array of temperature anomalies ordered by timestamp
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TempAnomaly'
  /api/printers/{deviceId}/jobs/{jobId}/pauses:
    get:
      tags: [Anomalies]
      summary: Get pause records for a job
      description: |
        Individual pause records with context about what caused the pause.
        Tracks both user-initiated pauses and error-induced pauses (when HMS errors are active).
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Array of pause records ordered by pause time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobPause'
  /api/printers/{deviceId}/anomalies:
    get:
      tags: [Anomalies]
      summary: Get temperature anomalies in a time window
      description: |
        Returns temperature anomalies for a printer within a time window, regardless of job.
        Useful for anomaly overlay when browsing arbitrary time ranges.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Start of time window (ISO 8601)
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: End of time window (ISO 8601)
      responses:
        '200':
          description: Array of temperature anomalies ordered by timestamp
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TempAnomaly'
        '400':
          description: Missing required query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/printers/{deviceId}/pauses:
    get:
      tags: [Anomalies]
      summary: Get job pauses in a time window
      description: |
        Returns pause records for a printer within a time window, regardless of job.
        Useful for pause band overlay when browsing arbitrary time ranges.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: Start of time window (ISO 8601)
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: End of time window (ISO 8601)
      responses:
        '200':
          description: Array of pause records ordered by pause time
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobPause'
        '400':
          description: Missing required query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Commands ───
  /api/printers/{deviceId}/command:
    post:
      tags: [Commands]
      summary: Send command to printer
      description: |
        Sends a command via cloud MQTT and waits up to 4 s for the printer's reply (matched by sequence_id).
        State-gated: pause needs RUNNING/PREPARE, resume needs PAUSE, stop/set_speed need an active print.
        Every attempt (including rejections) is recorded as a `command` event. Firmware with Bambu's
        authorization control may refuse unsigned commands — that surfaces as `ok:false` with the printer's reason.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command:
                  type: string
                  enum: [pause, resume, stop, set_speed]
                param:
                  description: Command parameter (e.g. speed level 1-4 for set_speed)
                  oneOf:
                    - type: string
                    - type: integer
      responses:
        '200':
          description: Command sent (check `acknowledged` / `ok` for the printer's answer)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, description: 'sent and not rejected by the printer' }
                  sent: { type: boolean }
                  acknowledged: { type: boolean, description: 'printer replied within 4 s' }
                  result: { type: string, nullable: true }
                  reason: { type: string, nullable: true }
                  outcome: { type: string }
        '409':
          description: Command not valid in the printer's current state
        '502':
          description: Could not publish (printer connection dropped)
        '503':
          description: Server not logged into BambuLab Cloud
        '400':
          description: Unknown command or invalid speed level
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Printer not found or not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Alerts ───
...
  /api/printers/{deviceId}/debug/mqtt:
    get:
      tags: [Debug]
      summary: Get raw MQTT merged state
      description: Returns the full merged MQTT state object for diagnostics. This is the raw data before extraction.
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Raw MQTT state
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                  connected:
                    type: boolean
                  mergedState:
                    type: object
                    description: Full raw MQTT merged state (printer-dependent structure)
        '404':
          description: Printer not found or not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Maintenance (BAM-39) ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Dashboard admin token (BAMBUZLE_ADMIN_TOKEN or <data dir>/admin-token)
    sessionCookie:
      type: apiKey
      in: cookie
      name: bambuzle_session
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
      description: Printer device ID
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: integer
      description: Print job ID
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    Sample:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: string
        job_id:
          type: integer
          nullable: true
        ts:
          type: string
          format: date-time
        bed_temp:
          type: number
          nullable: true
        bed_target:
          type: number
          nullable: true
        nozzle_temp:
          type: number
          nullable: true
        nozzle_target:
          type: number
          nullable: true
        nozzle2_temp:
          type: number
          nullable: true
        nozzle2_target:
          type: number
          nullable: true
        chamber_temp:
          type: number
          nullable: true
        part_fan_speed:
          type: integer
          nullable: true
        aux_fan_speed:
          type: integer
          nullable: true
        chamber_fan_speed:
          type: integer
          nullable: true
        progress:
          type: number
          nullable: true
        layer_num:
          type: integer
          nullable: true
        total_layers:
          type: integer
          nullable: true
        remaining_min:
          type: integer
          nullable: true
        gcode_state:
          type: string
        speed_level:
          type: integer
          nullable: true
        wifi_signal:
          type: integer
          nullable: true
    Event:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: string
        job_id:
          type: integer
          nullable: true
        ts:
          type: string
          format: date-time
        event_type:
          type: string
          enum: [state_change, hms_error, print_error, alert_fired, command]
        severity:
          type: string
          enum: [info, warning, error]
        code:
          type: string
          nullable: true
          description: HMS error code (for hms_error events)
        message:
          type: string
    PrintJob:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: string
        task_id:
          type: string
          nullable: true
        subtask_name:
          type: string
          nullable: true
        gcode_file:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        end_state:
          type: string
          nullable: true
          enum: [FINISH, FAILED, CANCELLED, IDLE]
        progress_pct:
          type: number
          nullable: true
        pause_count:
          type: integer
          description: Number of times the job was paused
        total_pause_sec:
          type: number
          description: Total seconds spent paused
        anomaly_count:
          type: integer
          description: Number of temperature anomalies detected during this job
        hms_codes:
          type: string
          nullable: true
          description: JSON array of HMS error codes seen during this job
        total_layers:
          type: integer
          nullable: true
          description: Total layer count for the print
        material:
          type: string
          nullable: true
          description: Active AMS tray filament type at job start (null if unknown / external spool)
        material_color:
          type: string
          nullable: true
          description: Active AMS tray colour (RRGGBBAA hex) at job start
        duration_sec:
          type: integer
          nullable: true
          description: Wall-clock job duration, set when the job ends
...
    JobStats:
      type: object
      properties:
        window:
          type: object
          properties:
            from:
              type: string
              description: UTC, 'YYYY-MM-DD HH:MM:SS'
            to:
              type: string
            printer:
              type: string
              nullable: true
        overall:
          $ref: '#/components/schemas/JobStatCounters'
        byPrinter:
          type: array
          items:
            allOf:
              - type: object
                properties:
                  deviceId:
                    type: string
                  name:
                    type: string
              - $ref: '#/components/schemas/JobStatCounters'
        byMaterial:
          type: array
          items:
            allOf:
              - type: object
                properties:
                  material:
                    type: string
                    description: Filament type (e.g. PLA), or 'Unknown'
              - $ref: '#/components/schemas/JobStatCounters'
        byDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              finished:
                type: integer
              failed:
                type: integer
              cancelled:
                type: integer
    LayerTransition:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: string
        job_id:
          type: integer
          nullable: true
        layer_num:
          type: integer
        ts:
          type: string
          format: date-time
        duration_sec:
          type: number
          nullable: true
          description: Seconds since previous layer transition (null for first layer)
        nozzle_temp:
          type: number
          nullable: true
        nozzle_target:
          type: number
          nullable: true
        bed_temp:
          type: number
          nullable: true
        bed_target:
          type: number
          nullable: true
        chamber_temp:
          type: number
          nullable: true
        speed_level:
          type: integer
          nullable: true
        progress:
          type: number
          nullable: true
    TempAnomaly:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: string
        job_id:
          type: integer
          nullable: true
        ts:
          type: string
          format: date-time
        sensor:
          type: string
          enum: [nozzle, nozzle2, bed, chamber]
        actual_temp:
          type: number
        target_temp:
          type: number
          nullable: true
        deviation:
          type: number
          nullable: true
          description: "Signed deviation from target (actual - target). Negative = undershoot, positive = overshoot."
        rate_of_change:
          type: number
          nullable: true
          description: Temperature change rate in degrees per second
        layer_num:
          type: integer
          nullable: true
        anomaly_type:
          type: string
          enum: [deviation, rate, both]
          description: What triggered this anomaly record
...